Completed skeleton to update statistics on robot

+++ src/edu/msoe/se2800/h4/StatsUpdateEvent.java
@@ -3,6 +3,6 @@ package edu.msoe.se2800.h4;
 public class StatsUpdateEvent {
     
     public int milliVolts;
-    public int velocity;
+    public double velocity;
 
 }
